Pull new version from dev

+++ assets/js/main.js
@@ -26,12 +26,12 @@ export class SpectrumData {
         }
     }
 }
-let spectrumData = new SpectrumData();
-let plot = new SpectrumPlot('plot');
-let raw = new RawData(1);
-let calClick = { a: false, b: false, c: false };
-let oldCalVals = { a: '', b: '', c: '' };
-let portsAvail = {};
+const spectrumData = new SpectrumData();
+const plot = new SpectrumPlot('plot');
+const raw = new RawData(1);
+const calClick = { a: false, b: false, c: false };
+const oldCalVals = { a: '', b: '', c: '' };
+const portsAvail = {};
 let refreshRate = 1000;
 let maxRecTimeEnabled = false;
 let maxRecTime = 1800000;
@@ -39,10 +39,10 @@ const REFRESH_META_TIME = 200;
 const CONSOLE_REFRESH = 200;
 let cpsValues = [];
 let isoListURL = 'assets/isotopes_energies_min.json';
-let isoList = {};
+const isoList = {};
 let checkNearIso = false;
 let maxDist = 100;
-const APP_VERSION = '2023-02-02';
+const APP_VERSION = '2023-02-03';
 let localStorageAvailable = false;
 let firstInstall = false;
 document.body.onload = async function () {
@@ -220,7 +220,7 @@ function importFile(input, background = false) {
     getFileData(input.files[0], background);
 }
 function getFileData(file, background = false) {
-    let reader = new FileReader();
+    const reader = new FileReader();
     const fileEnding = file.name.split('.')[1];
     reader.readAsText(file);
     reader.onload = async () => {
@@ -241,16 +241,28 @@ function getFileData(file, background = false) {
                 document.getElementById('add-notes').value = meta.notes;
                 startDate = new Date(meta.startTime);
                 endDate = new Date(meta.endTime);
-                if (!espectrum && !bgspectrum)
+                if (espectrum?.length && bgspectrum?.length) {
+                    spectrumData.data = espectrum;
+                    spectrumData.background = bgspectrum;
+                    spectrumData.dataTime = meta.dataMt * 1000;
+                    spectrumData.backgroundTime = meta.backgroundMt * 1000;
+                    if (meta.dataMt)
+                        spectrumData.dataCps = spectrumData.data.map(val => val / meta.dataMt);
+                    if (meta.backgroundMt)
+                        spectrumData.backgroundCps = spectrumData.background.map(val => val / meta.backgroundMt);
+                }
+                else if (!espectrum?.length && !bgspectrum?.length) {
                     popupNotification('file-error');
-                spectrumData.data = espectrum;
-                spectrumData.background = bgspectrum;
-                spectrumData.dataTime = meta.dataMt * 1000;
-                spectrumData.backgroundTime = meta.backgroundMt * 1000;
-                if (meta.dataMt)
-                    spectrumData.dataCps = spectrumData.data.map(val => val / meta.dataMt);
-                if (meta.backgroundMt)
-                    spectrumData.backgroundCps = spectrumData.background.map(val => val / meta.backgroundMt);
+                }
+                else {
+                    const fileData = espectrum?.length ? espectrum : bgspectrum;
+                    const fileDataTime = (espectrum?.length ? meta.dataMt : meta.backgroundMt) * 1000;
+                    const fileDataType = background ? 'background' : 'data';
+                    spectrumData[fileDataType] = fileData;
+                    spectrumData[`${fileDataType}Time`] = fileDataTime;
+                    if (fileDataTime)
+                        spectrumData[`${fileDataType}Cps`] = spectrumData[fileDataType].map(val => val / fileDataTime * 1000);
+                }
                 const importedCount = Object.values(coeff).filter(value => value !== 0).length;
                 if (importedCount >= 2) {
                     resetCal();
@@ -285,36 +297,52 @@ function getFileData(file, background = false) {
             document.getElementById('sample-weight').value = importData.sampleInfo?.weight?.toString() ?? '';
             document.getElementById('sample-vol').value = importData.sampleInfo?.volume?.toString() ?? '';
             document.getElementById('add-notes').value = importData.sampleInfo?.note ?? '';
-            if (importData.resultData.startTime) {
-                startDate = new Date(importData.resultData.startTime);
-                endDate = new Date(importData.resultData.endTime);
+            const resultData = importData.resultData;
+            if (resultData.startTime && resultData.endTime) {
+                startDate = new Date(resultData.startTime);
+                endDate = new Date(resultData.endTime);
             }
-            const localKeys = ['data', 'background'];
-            const importKeys = ['energySpectrum', 'backgroundEnergySpectrum'];
-            for (const i in localKeys) {
-                const newKey = importKeys[i];
-                if (newKey in importData.resultData) {
-                    spectrumData[localKeys[i]] = importData.resultData[newKey].spectrum;
-                    if ('measurementTime' in importData.resultData[newKey] && importData.resultData[newKey].measurementTime > 0) {
-                        spectrumData[`${localKeys[i]}Time`] = importData.resultData[newKey].measurementTime * 1000;
-                        spectrumData[`${localKeys[i]}Cps`] = spectrumData[localKeys[i]].map(val => val / importData.resultData[newKey].measurementTime);
-                    }
-                    if ('energyCalibration' in importData.resultData[newKey]) {
-                        const coeffArray = importData.resultData[newKey].energyCalibration.coefficients;
-                        const numCoeff = importData.resultData[newKey].energyCalibration.polynomialOrder;
-                        resetCal();
-                        for (const index in coeffArray) {
-                            plot.calibration.coeff[`c${numCoeff - parseInt(index) + 1}`] = coeffArray[index];
-                        }
-                        plot.calibration.imported = true;
-                        displayCoeffs();
-                        const calSettings = document.getElementsByClassName('cal-setting');
-                        for (const element of calSettings) {
-                            element.disabled = true;
-                        }
-                        addImportLabel();
-                    }
+            const espectrum = resultData.energySpectrum;
+            const bgspectrum = resultData.backgroundEnergySpectrum;
+            if (espectrum && bgspectrum) {
+                spectrumData.data = espectrum.spectrum;
+                spectrumData.background = bgspectrum.spectrum;
+                const eMeasurementTime = espectrum.measurementTime;
+                if (eMeasurementTime) {
+                    spectrumData.dataTime = eMeasurementTime * 1000;
+                    spectrumData.dataCps = spectrumData.data.map(val => val / eMeasurementTime);
                 }
+                const bgMeasurementTime = bgspectrum.measurementTime;
+                if (bgMeasurementTime) {
+                    spectrumData.backgroundTime = bgMeasurementTime * 1000;
+                    spectrumData.backgroundCps = spectrumData.background.map(val => val / bgMeasurementTime);
+                }
+            }
+            else {
+                const dataObj = espectrum ?? bgspectrum;
+                const fileData = dataObj?.spectrum ?? [];
+                const fileDataTime = (dataObj?.measurementTime ?? 1) * 1000;
+                const fileDataType = background ? 'background' : 'data';
+                spectrumData[fileDataType] = fileData;
+                spectrumData[`${fileDataType}Time`] = fileDataTime;
+                if (fileDataTime)
+                    spectrumData[`${fileDataType}Cps`] = spectrumData[fileDataType].map(val => val / fileDataTime * 1000);
+            }
+            const calDataObj = (espectrum ?? bgspectrum)?.energyCalibration;
+            if (calDataObj) {
+                const coeffArray = calDataObj.coefficients;
+                const numCoeff = calDataObj.polynomialOrder;
+                resetCal();
+                for (const index in coeffArray) {
+                    plot.calibration.coeff[`c${numCoeff - parseInt(index) + 1}`] = coeffArray[index];
+                }
+                plot.calibration.imported = true;
+                displayCoeffs();
+                const calSettings = document.getElementsByClassName('cal-setting');
+                for (const element of calSettings) {
+                    element.disabled = true;
+                }
+                addImportLabel();
             }
         }
         else if (background) {
@@ -327,7 +355,7 @@ function getFileData(file, background = false) {
         }
         updateSpectrumCounts();
         updateSpectrumTime();
-        if (!(spectrumData.background.length === spectrumData.data.length || !spectrumData.data.length || !spectrumData.background.length)) {
+        if (spectrumData.background.length !== spectrumData.data.length && spectrumData.data.length && spectrumData.background.length) {
             popupNotification('data-error');
             removeFile(background ? 'background' : 'data');
         }
@@ -353,8 +381,8 @@ function removeFile(id) {
     spectrumData[id] = [];
     spectrumData[`${id}Time`] = 0;
     document.getElementById(id).value = '';
-    document.getElementById('total-spec-cts').innerText = spectrumData.getTotalCounts('data').toString();
-    document.getElementById('total-bg-cts').innerText = spectrumData.getTotalCounts('background').toString();
+    updateSpectrumCounts();
+    updateSpectrumTime();
     document.getElementById(id + '-icon').classList.add('d-none');
     plot.resetPlot(spectrumData);
     bindPlotEvents();
@@ -409,7 +437,7 @@ function resetPlot() {
 document.getElementById('xAxis').onclick = event => changeAxis(event.target);
 document.getElementById('yAxis').onclick = event => changeAxis(event.target);
 function changeAxis(button) {
-    let id = button.id;
+    const id = button.id;
     if (plot[id] === 'linear') {
         plot[id] = 'log';
         button.innerText = 'Log';
@@ -485,13 +513,13 @@ function toggleCal(enabled) {
     button.innerHTML = enabled ? '<i class="fa-solid fa-rotate-left"></i> Reset' : '<i class="fa-solid fa-check"></i> Calibrate';
     if (enabled) {
         if (!plot.calibration.imported) {
-            let readoutArray = [
+            const readoutArray = [
                 [document.getElementById('adc-a').value, document.getElementById('cal-a').value],
                 [document.getElementById('adc-b').value, document.getElementById('cal-b').value],
                 [document.getElementById('adc-c').value, document.getElementById('cal-c').value]
             ];
             let invalid = 0;
-            let validArray = [];
+            const validArray = [];
             for (const pair of readoutArray) {
                 const float1 = parseFloat(pair[0]);
                 const float2 = parseFloat(pair[1]);
@@ -579,13 +607,13 @@ function importCalButton(input) {
     importCal(input.files[0]);
 }
 function importCal(file) {
-    let reader = new FileReader();
+    const reader = new FileReader();
     reader.readAsText(file);
     reader.onload = () => {
         try {
             const result = reader.result.trim();
             const obj = JSON.parse(result);
-            let readoutArray = [
+            const readoutArray = [
                 document.getElementById('adc-a'),
                 document.getElementById('cal-a'),
                 document.getElementById('adc-b'),
@@ -669,45 +697,45 @@ function downloadCal() {
 }
 function makeXMLSpectrum(type, name) {
     const root = document.createElementNS(null, (type === 'data') ? 'EnergySpectrum' : 'BackgroundEnergySpectrum');
-    let noc = document.createElementNS(null, 'NumberOfChannels');
+    const noc = document.createElementNS(null, 'NumberOfChannels');
     noc.textContent = spectrumData[type].length.toString();
     root.appendChild(noc);
-    let sn = document.createElementNS(null, 'SpectrumName');
+    const sn = document.createElementNS(null, 'SpectrumName');
     sn.textContent = name;
     root.appendChild(sn);
     if (plot.calibration.enabled) {
-        let ec = document.createElementNS(null, 'EnergyCalibration');
+        const ec = document.createElementNS(null, 'EnergyCalibration');
         root.appendChild(ec);
-        let c = document.createElementNS(null, 'Coefficients');
-        let coeffs = [];
+        const c = document.createElementNS(null, 'Coefficients');
+        const coeffs = [];
         const coeffObj = plot.calibration.coeff;
         for (const index in coeffObj) {
             coeffs.push(coeffObj[index]);
         }
         const coeffsRev = coeffs.reverse();
         for (const val of coeffsRev) {
-            let coeff = document.createElementNS(null, 'Coefficient');
+            const coeff = document.createElementNS(null, 'Coefficient');
             coeff.textContent = val.toString();
             c.appendChild(coeff);
         }
         ec.appendChild(c);
-        let po = document.createElementNS(null, 'PolynomialOrder');
+        const po = document.createElementNS(null, 'PolynomialOrder');
         po.textContent = (2).toString();
         ec.appendChild(po);
     }
-    let tpc = document.createElementNS(null, 'TotalPulseCount');
+    const tpc = document.createElementNS(null, 'TotalPulseCount');
     tpc.textContent = spectrumData.getTotalCounts(type).toString();
     root.appendChild(tpc);
-    let vpc = document.createElementNS(null, 'ValidPulseCount');
+    const vpc = document.createElementNS(null, 'ValidPulseCount');
     vpc.textContent = tpc.textContent;
     root.appendChild(vpc);
-    let mt = document.createElementNS(null, 'MeasurementTime');
+    const mt = document.createElementNS(null, 'MeasurementTime');
     mt.textContent = (Math.round(spectrumData[`${type}Time`] / 1000)).toString();
     root.appendChild(mt);
-    let s = document.createElementNS(null, 'Spectrum');
+    const s = document.createElementNS(null, 'Spectrum');
     root.appendChild(s);
     for (const datapoint of spectrumData[type]) {
-        let d = document.createElementNS(null, 'DataPoint');
+        const d = document.createElementNS(null, 'DataPoint');
         d.textContent = datapoint.toString();
         s.appendChild(d);
     }
@@ -719,7 +747,7 @@ function downloadXML() {
     const formatVersion = 230124;
     const spectrumName = getDateStringMin() + ' Energy Spectrum';
     const backgroundName = getDateStringMin() + ' Background Energy Spectrum';
-    let doc = document.implementation.createDocument(null, "ResultDataFile");
+    const doc = document.implementation.createDocument(null, "ResultDataFile");
     const pi = doc.createProcessingInstruction('xml', 'version="1.0" encoding="UTF-8"');
     doc.insertBefore(pi, doc.firstChild);
     const root = doc.documentElement;
@@ -791,7 +819,7 @@ function downloadXML() {
     download(filename, new XMLSerializer().serializeToString(doc));
 }
 function makeJSONSpectrum(type) {
-    let spec = {
+    const spec = {
         'numberOfChannels': spectrumData[type].length,
         'validPulseCount': spectrumData.getTotalCounts(type),
         'measurementTime': 0,
@@ -799,7 +827,7 @@ function makeJSONSpectrum(type) {
     };
     spec.measurementTime = Math.round(spectrumData[`${type}Time`] / 1000);
     if (plot.calibration.enabled) {
-        let calObj = {
+        const calObj = {
             'polynomialOrder': 0,
             'coefficients': []
         };
@@ -812,7 +840,7 @@ function makeJSONSpectrum(type) {
 document.getElementById('npes-export-btn').onclick = () => downloadNPES();
 function downloadNPES() {
     const filename = `spectrum_${getDateString()}.json`;
-    let data = {
+    const data = {
         'schemaVersion': 'NPESv1',
         'deviceData': {
             'softwareName': 'Gamma MCA, ' + APP_VERSION,
@@ -915,7 +943,7 @@ async function loadIsotopes(reload = false) {
             tableElement.innerHTML = '';
             plot.clearAnnos();
             plot.updatePlot(spectrumData);
-            let intKeys = Object.keys(json);
+            const intKeys = Object.keys(json);
             intKeys.sort((a, b) => parseFloat(a) - parseFloat(b));
             let index = 0;
             for (const key of intKeys) {
@@ -971,7 +999,7 @@ async function closestIso(value) {
     if (!isNaN(energyVal))
         plot.toggleLine(energyVal, Object.keys(prevIso)[0], false);
     if (energy && name) {
-        let newIso = {};
+        const newIso = {};
         newIso[energy] = name;
         if (prevIso !== newIso)
             prevIso = newIso;
@@ -1044,10 +1072,7 @@ function loadSettingsDefault() {
     document.getElementById('custom-ser-refresh').value = (refreshRate / 1000).toString();
     document.getElementById('custom-ser-buffer').value = SerialManager.maxSize.toString();
     document.getElementById('custom-ser-adc').value = SerialManager.adcChannels.toString();
-    const autoStop = document.getElementById('ser-limit');
-    autoStop.value = (maxRecTime / 1000).toString();
-    autoStop.disabled = !maxRecTimeEnabled;
-    document.getElementById('ser-limit-btn').disabled = !maxRecTimeEnabled;
+    document.getElementById('ser-limit').value = (maxRecTime / 1000).toString();
     document.getElementById('toggle-time-limit').checked = maxRecTimeEnabled;
     document.getElementById('iso-hover-prox').value = maxDist.toString();
     document.getElementById('custom-baud').value = SerialManager.serOptions.baudRate.toString();
@@ -1178,8 +1203,6 @@ function changeSettings(name, element) {
             break;
         case 'timeLimitBool':
             boolVal = element.checked;
-            document.getElementById('ser-limit').disabled = !boolVal;
-            document.getElementById('ser-limit-btn').disabled = !boolVal;
             maxRecTimeEnabled = boolVal;
             saveJSON(name, boolVal);
             break;
@@ -1269,7 +1292,6 @@ function serialConnect() {
     listSerial();
     popupNotification('serial-connect');
 }
-;
 function serialDisconnect(event) {
     for (const key in portsAvail) {
         if (portsAvail[key] == event.target) {
@@ -1282,7 +1304,6 @@ function serialDisconnect(event) {
     listSerial();
     popupNotification('serial-disconnect');
 }
-;
 document.getElementById('serial-list-btn').onclick = () => listSerial();
 async function listSerial() {
     const portSelector = document.getElementById('port-selector');
